rename walk.ts -> manifest.ts
