+++ src/utils/manifest.ts
